timer: various timer fixes
1) when no next timer, pass ~0 to platSetTimerAlarm
2) subtract off how long it takes to wakeup when preparing to sleep
3) change period to uin64_t so it is the same size as expires

+++ firmware/src/timer.c
@@ -12,7 +12,7 @@
 
 struct Timer {
     uint64_t      expires; /* time of next expiration */
-    uint32_t      period;  /* 0 for oneshot */
+    uint64_t      period;  /* 0 for oneshot */
     uint32_t      id;      /* 0 for disabled */
     TimTimerCbkF  cbk;
     uint32_t      jitterPpm;
